# Optimize generated code size (Scala 3)