--- a/core/src/main/scala-3/caliban/schema/macros/Macros.scala
+++ b/core/src/main/scala-3/caliban/schema/macros/Macros.scala
@@ -1,6 +1,6 @@
 package caliban.schema.macros
 
-import caliban.schema.Annotations.{GQLInterface, GQLUnion, GQLValueType}
+import caliban.schema.Annotations.{ GQLInterface, GQLUnion, GQLValueType }
 
 import scala.quoted.*