As discussed [on Discord](https://discord.com/channels/629491597070827530/633200096393166868/1075604452641361991), the Scala 3 generated code is currently quite large, mostly due to boilerplate code being inlined for every derived type.

~In this PR, we use macros along with `inline if` so that we don't generate redundant code for branches we already know during compile-time that won't be called during runtime.~

In this PR, we move the creation of the `Schema[R, A]` object into a private method so that the code is not generated for each derived type

I tested these changes on a medium size application, and the application code size went down from 2.7MB to 1.3MB. The number of classes generated via Caliban also decreased from ~900 to ~300. This makes the Scala 3 generated code roughly the same size as the Scala 2 counterpart

TODO:

* [x] Optimise ArgBuilder
* [x] ~Figure out how to refactor the `recurse` and `isEnum` into a macro~ Rewrite `recurse` to be tail recursive (generated code becomes a flat list)