@kyri-petrou 2023-02-16
I can't seem to figure out how to rewrite [this](https://github.com/ghostdogpr/caliban/pull/1607/files#diff-88ee5242eb50f55a96a2ae7a676d4b116c93fe6f9653dbe2ed8afa6193dbd16cL34-L42) into a macro due to it being recursive. I haven't worked with macros before though so I'm probably missing something super easy / obvious: Any pointers are much appreciated 😄

Observing the generated scala code (using the `-Xprint:erasure` compiler flag) this is causing quite a fair bit of code to be generated. If we manage to convert that into a macro, `isEnum` I think it'll cut down the size of the generated code by another 20-30%, which will bring it fairly close to the levels of Magnolia. Unlike magnolia, for most parts we're not just creating a super generic template method but rather get the compiler to strip out the parts we don't need, which I think might be better for runtime performance