@ghostdogpr 2023-02-16
Ugh, I wanted to avoid macros as much as possible because I can't maintain them 😂

So the private def functions that don't have the inline keyword are inlined? I didn't expect that 🤔