@kyri-petrou 2023-02-16
> Ugh, I wanted to avoid macros as much as possible because I can't maintain them 😂

Hah I'm totally the same 🥲. This is actually the first time I touched macros... ever. [The ones that I added](https://github.com/ghostdogpr/caliban/pull/1607/files#diff-9f8fb5fe145689da082577535c038e431584b0873871c7bf91ba12e97b4d75cdR67-R108) seem quite trivial, they're just a typecheck really. I'm happy to remove them and capitalise more on private methods (see more below) if you think they're going to make it more difficult maintaining the library

> So the private def functions that don't have the inline keyword are inlined? I didn't expect that 🤔

The private functions are not inlined no, that's why I moved [these](https://github.com/ghostdogpr/caliban/pull/1607/files#diff-88ee5242eb50f55a96a2ae7a676d4b116c93fe6f9653dbe2ed8afa6193dbd16cR209-R261) into private functions to avoid the `.map`, `.filter` and all that being generated for each schema (TIL that these methods generate quite a bit of bytecode). The main issue is with [this method](https://github.com/ghostdogpr/caliban/blob/series/2.x/core/src/main/scala-3/caliban/schema/SchemaDerivation.scala#L34-L42), which inlines a lot of code for nested types. If we could refactor to a macro, that would reduce the size of the generated code quite a bit (and improve compile time as well). Having said that, I think we should only do that if there's a simple-ish way to do it, otherwise adding too much complexity for the sake of performance is not worth it IMO