@kyri-petrou 2023-02-17
Benchmarks for old vs new implementation. There is a small-ish improvement, most of it which seems to come from GC (which I guess is expected). I'll make the changes discussed on discord and rerun the tests

### Old

```
[info] GraphQLBenchmarks.fragmentsCaliban                                         thrpt    5        59.247 ±    1.307   ops/s
[info] GraphQLBenchmarks.fragmentsCaliban:·gc.alloc.rate                          thrpt    5      3498.826 ±   77.144  MB/sec
[info] GraphQLBenchmarks.fragmentsCaliban:·gc.alloc.rate.norm                     thrpt    5  61926024.267 ± 1015.661    B/op
[info] GraphQLBenchmarks.fragmentsCaliban:·gc.count                               thrpt    5       210.000             counts
[info] GraphQLBenchmarks.fragmentsCaliban:·gc.time                                thrpt    5       213.000                 ms

[info] GraphQLBenchmarks.introspectCaliban                                        thrpt    5      3175.157 ±   83.557   ops/s
[info] GraphQLBenchmarks.introspectCaliban:·gc.alloc.rate                         thrpt    5      3229.301 ±   84.998  MB/sec
[info] GraphQLBenchmarks.introspectCaliban:·gc.alloc.rate.norm                    thrpt    5   1066508.796 ±   66.851    B/op
[info] GraphQLBenchmarks.introspectCaliban:·gc.count                              thrpt    5       324.000             counts
[info] GraphQLBenchmarks.introspectCaliban:·gc.time                               thrpt    5       144.000                 ms

[info] execution.NestedZQueryBenchmark.deepParallelQuery1000                      thrpt    5       107.663 ±    1.895   ops/s
[info] execution.NestedZQueryBenchmark.deepParallelQuery1000:·gc.alloc.rate       thrpt    5      7016.640 ±  123.402  MB/sec
[info] execution.NestedZQueryBenchmark.deepParallelQuery1000:·gc.alloc.rate.norm  thrpt    5  68341118.284 ± 1784.521    B/op
[info] execution.NestedZQueryBenchmark.deepParallelQuery1000:·gc.count            thrpt    5       115.000             counts
[info] execution.NestedZQueryBenchmark.deepParallelQuery1000:·gc.time             thrpt    5       683.000                 ms
```

### New (with macros)

```
[info] GraphQLBenchmarks.fragmentsCaliban                                         thrpt    5        62.621 ±    1.534   ops/s
[info] GraphQLBenchmarks.fragmentsCaliban:·gc.alloc.rate                          thrpt    5      3536.092 ±   86.562  MB/sec
[info] GraphQLBenchmarks.fragmentsCaliban:·gc.alloc.rate.norm                     thrpt    5  59213344.178 ±  817.945    B/op
[info] GraphQLBenchmarks.fragmentsCaliban:·gc.count                               thrpt    5       186.000             counts
[info] GraphQLBenchmarks.fragmentsCaliban:·gc.time                                thrpt    5       195.000                 ms

[info] GraphQLBenchmarks.introspectCaliban                                        thrpt    5      3219.513 ±    8.513   ops/s
[info] GraphQLBenchmarks.introspectCaliban:·gc.alloc.rate                         thrpt    5      3166.384 ±    8.356  MB/sec
[info] GraphQLBenchmarks.introspectCaliban:·gc.alloc.rate.norm                    thrpt    5   1031309.281 ±   63.995    B/op
[info] GraphQLBenchmarks.introspectCaliban:·gc.count                              thrpt    5       259.000             counts
[info] GraphQLBenchmarks.introspectCaliban:·gc.time                               thrpt    5       125.000                 ms

[info] execution.NestedZQueryBenchmark.deepParallelQuery1000                      thrpt    5       109.888 ±    2.408   ops/s
[info] execution.NestedZQueryBenchmark.deepParallelQuery1000:·gc.alloc.rate       thrpt    5      6819.714 ±  149.341  MB/sec
[info] execution.NestedZQueryBenchmark.deepParallelQuery1000:·gc.alloc.rate.norm  thrpt    5  65078023.859 ± 1232.956    B/op
[info] execution.NestedZQueryBenchmark.deepParallelQuery1000:·gc.count            thrpt    5       124.000             counts
[info] execution.NestedZQueryBenchmark.deepParallelQuery1000:·gc.time             thrpt    5       578.000                 ms
```