@kyri-petrou 2023-02-17
@ghostdogpr  Okay so it seems the benefits the macros are providing is marginal. I removed them and introduced `makeXXSchema` methods instead.

With these changes, the Scala 3 codegen is now about the same as Scala 2 :)